Optimized Imports and restructured CalendarTUI
CalendarTUI is now more readable and uses better representations of time. Also it should be able to display Tasks, although due to compile issues I am currently not able to test the changes made. There currently seems to be an issue with the recognition of the packaged model, view and controller. From all of these packages no imports are possible.